Outputters get an instance of Report now
