+++ psireporter/plugin.py
@@ -25,29 +25,3 @@ class ReporterPlugin(PluginMeta):
     def __init__(cls, name, bases, namespaces):
         super(ReporterPlugin, cls).__init__("reporters", name, bases, namespaces)
 
-
-class Report():
-
-    def __init__(self, *args, **kwargs):
-        self._id = kwargs.get('id', None)
-
-        if self._id is None:
-            self._id = str(uuid.uuid1())
-
-        self._message = kwargs.get('message', None)
-
-        d = datetime.utcnow()
-        unixtime = calendar.timegm(d.utctimetuple())
-        self._timestamp = unixtime
-
-    @property
-    def id(self):
-        return self._id
-
-    @property
-    def message(self):
-        return self._message
-
-    @property
-    def timestamp(self):
-        return self._timestamp
